requirements.txt

apache-airflow
chromadb

ollama
